requirement.txt
langchain 
wikipedia 
langchain-community 
langchain-google-genai 
google-generativeai
python-dotenv
pydantic
duckduckgo-search